+++ .eslintignore
@@ -1,2 +1,3 @@
 **/*.test.js
+config/setupTests.js
 scripts/*
