+++ config/setupTests.js
@@ -11,3 +11,19 @@ global.render = render;
 global.mount = mount;
 global.React = React;
 global.fetch = fetch;
+
+window.insights = {
+    chrome: {
+        getUserPermissions: () => Promise.resolve()
+    },
+    loadInventory: () => Promise.resolve({
+        inventoryConnector: (args) => {
+            const InventoryTable = () => <div>A mock passed!</div>;
+            return ({ InventoryTable });
+        },
+        mergeWithEntities: () => { },
+        mergeWithDetail: () => { }
+    })
+};
+
+jest.useFakeTimers('modern').setSystemTime(new Date('2020-01-01').getTime());
