+++ src/Components/PresentationalComponents/CVEDetailsPageSidebar/SnippetWithHeaderAndPopover.js
@@ -3,27 +3,24 @@ import propTypes from 'prop-types';
 import React from 'react';
 import Label from '../Snippets/Label';
 
-const SnippetWithHeaderAndPopover = props => {
-    const { title, value, content } = props;
-
-    return (
-        <Popover id="popover" bodyContent={content} headerContent="" position="bottom"
-            appendTo={document.querySelector('.vulnerability')}>
-            <Stack className="popover-content">
-                <StackItem>
-                    <Label isLarge>{title}</Label>
-                </StackItem>
-                <StackItem>
-                    <a>{value}</a>
-                </StackItem>
-            </Stack>
-        </Popover>
-    );
-};
+const SnippetWithHeaderAndPopover = ({ title, value, content }) => (
+    <Popover id="popover" bodyContent={content} headerContent="" position="bottom"
+        appendTo={document.querySelector('.vulnerability')}>
+        <Stack className="popover-content">
+            <StackItem>
+                <Label isLarge>{title}</Label>
+            </StackItem>
+            <StackItem>
+                <a>{value}</a>
+            </StackItem>
+        </Stack>
+    </Popover>
+);
 
 SnippetWithHeaderAndPopover.propTypes = {
     title: propTypes.any,
     value: propTypes.any,
     content: propTypes.any
 };
+
 export default SnippetWithHeaderAndPopover;
