+++ src/Components/PresentationalComponents/CvssVector/CvssVector.js
@@ -24,8 +24,7 @@ export const parseVector = (vector, namesMapping) => {
     return res;
 };
 
-const CvssVector = props => {
-    const { cvss2_metrics: CVSS2Vector, cvss3_metrics: CVSS3Vector, intl } = props;
+const CvssVector = ({ cvss2_metrics: CVSS2Vector, cvss3_metrics: CVSS3Vector, score, intl }) => {
     const cvssVer = (CVSS3Vector && 'CVSS 3.0') || (CVSS2Vector && 'CVSS 2.0') || 'CVSS 3.0';
     const cvssVector = CVSS3Vector || CVSS2Vector;
     const namesMapping = (CVSS3Vector && V3names) || (CVSS2Vector && V2names);
@@ -73,7 +72,7 @@ const CvssVector = props => {
                     </Popover>
 
                     <WithLoader isLoading={context.isLoading} style={{ width: '320px' }}>
-                        <span className="pf-u-mr-md">{props.score}</span>
+                        <span className="pf-u-mr-md">{score}</span>
                         <span id="cvss-vector-content">
                             {intl.formatMessage(messages.vector) + ': '}
                             {cvssVector?.substring(cvssVector.indexOf('/') + 1) || notAvailable}
