+++ src/Components/PresentationalComponents/StaticPages/NoAccessPage.test.js
@@ -9,8 +9,6 @@ import { intl } from '../../../Utilities/IntlProvider';
 const mockStore = configureStore([store => next => action => {}]);
 let store = mockStore({ });
 
-window.insights = { };
-
 describe('NoAccessPage component', () => {
     it('Should match snapshot', () => {
         const wrapper = mountWithIntl(
@@ -43,6 +41,6 @@ describe('NoAccessPage component', () => {
                 </Router>
             </Provider>);        
 
-    expect(wrapperWithReferer.find('button').text()).toEqual('Return to previous page');
+        expect(wrapperWithReferer.find('button').text()).toEqual('Return to previous page');
     })
 });
