+++ src/Components/PresentationalComponents/StaticPages/UpgradePage.test.js
@@ -9,8 +9,6 @@ import { intl } from '../../../Utilities/IntlProvider';
 const mockStore = configureStore([store => next => action => {}]);
 let store = mockStore({ });
 
-window.insights = { };
-
 describe('UpgradePage component', () => {
     it('Should match snapshot', () => {
         const wrapper = mountWithIntl(
