+++ src/Components/SmartComponents/CVEDetailsPage/CveDetailsPage.test.js
@@ -70,16 +70,7 @@ const customMiddleWare = store => next => action => {
 const mockStore = configureStore([customMiddleWare]);
 const store = mockStore(initialState);
 const props = { match: { params: { cve: 'testCve' } } };
-window.insights = {
-    loadInventory: (args) => new Promise.resolve({
-        inventoryConnector: (args) => {
-            const InventoryTable = ({ testProp }) => <div>A mock with '{testProp}' passed!</div>;
-            return ({ InventoryTable });
-        },
-        mergeWithEntities: () => { },
-        mergeWithDetail: () => { }
-    })
-};
+
 describe('CVE details:', () => {
     it('Should render match snapshot', () => {
         const wrapper = shallow(
