+++ src/Components/SmartComponents/CVEs/CVEs.test.js
@@ -92,6 +92,7 @@ beforeEach(() => {
     useSelector.mockImplementation(callback => {
         return callback({ CVEsStore: initialState });
     });
+
     wrapper = mountWithIntl(
         <Provider store={store}>
             <Router><CVEs /></Router>
