+++ src/Components/SmartComponents/CVEs/CVEsTable.test.js
@@ -7,13 +7,10 @@ import { handleSortColumn } from '../../../Helpers/MiscHelper';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { mountWithIntl } from '../../../Helpers/MiscHelper';
 
-
-jest.mock('../../../Helpers/MiscHelper', () => (
-    {
-      ...(jest.requireActual('../../../Helpers/MiscHelper')),
-      handleSortColumn: jest.fn(),
-    }
-));
+jest.mock('../../../Helpers/MiscHelper', () => ({
+    ...(jest.requireActual('../../../Helpers/MiscHelper')),
+    handleSortColumn: jest.fn(),
+}));
 
 const mockContext = {
     cves: createCveListByAccount({ payload: {
