+++ src/Components/SmartComponents/CVEs/CVEsTableToolbar.js
@@ -27,18 +27,16 @@ import {
 } from '../../../Helpers/TableToolbarHelper';
 import { CVES_DEFAULT_FILTERS, CVES_FILTER_PARAMS } from '../../../Helpers/constants';
 
-const CVEsTableToolbarWithContext = (props) => {
+const CVEsTableToolbarWithContext = ({ context, canEditStatusOrBusinessRisk, canExport, intl }) => {
     const [exportPDF, setExportPDF] = useState(false);
 
     const onExpandAllClick = () => {
-        const { context } = props;
         const { cves, methods, isAllExpanded } = context;
 
         const expandedRows = !isAllExpanded ? cves.data.filter(cve => cve.id).map(cve => cve.id) : [];
         methods.openCves(expandedRows);
     };
 
-    const { context, canEditStatusOrBusinessRisk, canExport } = props;
     const { cves, params, methods, selectedCves, isAllExpanded, selectedRowsRawData } = context;
     const { filter } = params;
     const selectedCvesCount = selectedCves && selectedCves.length;
@@ -55,7 +53,7 @@ const CVEsTableToolbarWithContext = (props) => {
         '', // #NOTE empty intentionally, Remediation holder
         ...canEditStatusOrBusinessRisk ? [
             {
-                label: props.intl.formatMessage(messages.editBusinessRisk),
+                label: intl.formatMessage(messages.editBusinessRisk),
                 onClick: () => methods.showBusinessRiskModal(selectedRowsRawData.map(({ id, attributes }) => ({
                     id,
                     business_risk_id: attributes.business_risk_id,
@@ -64,7 +62,7 @@ const CVEsTableToolbarWithContext = (props) => {
                 props: { isDisabled: !selectedCvesCount }
             },
             {
-                label: props.intl.formatMessage(messages.editStatus),
+                label: intl.formatMessage(messages.editStatus),
                 onClick: () => methods.showStatusModal(selectedRowsRawData.map(({ id, attributes }) => ({
                     id,
                     exposed_systems_count: attributes.systems_affected,
@@ -75,7 +73,7 @@ const CVEsTableToolbarWithContext = (props) => {
             }
         ] : [],
         {
-            label: props.intl.formatMessage(messages.columnManagementModalTitle),
+            label: intl.formatMessage(messages.columnManagementModalTitle),
             onClick: () => methods.setColumnModalOpen(true)
         }
     ];
@@ -120,7 +118,7 @@ const CVEsTableToolbarWithContext = (props) => {
                 activeFiltersConfig={{
                     filters: buildActiveFilters(params),
                     onDelete: (_, chips, reset) => removeFilters(chips, methods.apply, reset, CVES_DEFAULT_FILTERS),
-                    deleteTitle: props.intl.formatMessage(messages.resetFilters),
+                    deleteTitle: intl.formatMessage(messages.resetFilters),
                     showDeleteButton: !isFilterInDefaultState(params, CVES_DEFAULT_FILTERS, CVES_FILTER_PARAMS)
                 }}
                 exportConfig={canExport && {
@@ -165,4 +163,5 @@ const CVEsTableToolbar = props => (
         {context => <CVEsTableToolbarWithContext context={context} {...props} />}
     </CVETableContext.Consumer>
 );
+
 export default injectIntl(CVEsTableToolbar);
