+++ src/Components/SmartComponents/CVEs/CVEsTableToolbar.test.js
@@ -8,14 +8,12 @@ import {
     removeFilters
 } from '../../../Helpers/TableToolbarHelper';
 
-jest.mock('../../../Helpers/TableToolbarHelper', () => (
-    {
-      ...(jest.requireActual('../../../Helpers/TableToolbarHelper')),
-      handleChangePage: jest.fn(),
-      handleSetPageSize: jest.fn(),
-      removeFilters: jest.fn()
-    }
-));
+jest.mock('../../../Helpers/TableToolbarHelper', () => ({
+    ...(jest.requireActual('../../../Helpers/TableToolbarHelper')),
+    handleChangePage: jest.fn(),
+    handleSetPageSize: jest.fn(),
+    removeFilters: jest.fn()
+}));
 
 jest.mock('../../../Store/Actions/Actions', () => ({
     ...jest.requireActual('../../../Store/Actions/Actions'),
