+++ src/Components/SmartComponents/Dashbar/Dashbar.test.js
@@ -5,12 +5,10 @@ import { mountWithIntl } from '../../../Helpers/MiscHelper';
 import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
 import { BrowserRouter as Router } from 'react-router-dom';
-window.insights = {};
 
 jest.mock("react-redux", () => ({
     ...jest.requireActual("react-redux"),
-    useSelector: jest.fn(),
-    useDispatch: jest.fn()
+    useSelector: jest.fn()
 }));
 
 const mockStore = configureStore([store => next => action => { }]);
