+++ src/Components/SmartComponents/LandingPage/LandingPage.test.js
@@ -6,8 +6,7 @@ import { shallowWithIntl } from '../../../Helpers/MiscHelper';
 
 jest.mock("react-redux", () => ({
     ...jest.requireActual("react-redux"),
-    useSelector: jest.fn(),
-    useDispatch: jest.fn()
+    useSelector: jest.fn()
 }));
 
 describe('CVEs: ', () => {
