+++ src/Components/SmartComponents/Modals/BusinessRiskModal.test.js
@@ -6,6 +6,7 @@ import ReducerRegistry from '../../../Utilities/ReducerRegistry';
 import {intl} from '../../../Utilities/IntlProvider';
 import { mountWithIntl } from '../../../Helpers/MiscHelper';
 import { act } from 'react-dom/test-utils';
+
 const MockStore = ({ children }) => <Provider store={ReducerRegistry.getStore()}>{children}</Provider>;
 
 describe('BusinessRiskModal', () => {
