+++ src/Components/SmartComponents/Modals/ColumnManagementModal.test.js
@@ -34,17 +34,18 @@ export const SYSTEM_DETAILS_HEADER = [
         transforms: [sortable, wrappable],
         isShownByDefault: false
     },
-
 ];
 
 let wrapper;
 beforeEach(() => {
     wrapper = mountWithIntl(
         <Router>
-            <ColumnManagementModal appliedColumns={SYSTEM_DETAILS_HEADER}
-                                   applyColumns={null}
-                                   isModalOpen={true}
-                                   setModalOpen={null}/>
+            <ColumnManagementModal
+                appliedColumns={SYSTEM_DETAILS_HEADER}
+                applyColumns={null}
+                isModalOpen={true}
+                setModalOpen={null}
+            />
         </Router>
     );
 });
