+++ src/Components/SmartComponents/Modals/ReportConfigModal.test.js
@@ -7,9 +7,6 @@ import { useSelector } from 'react-redux';
 import ReportConfigModal from './ReportConfigModal';
 import { act } from 'react-dom/test-utils';
 import { CVE_REPORT_FILTERS, DEFAULT_FILTER_DATA, CUSTOM_REPORT_DEFAULT_SORT } from '../../../Helpers/constants';
-import DownloadCVEsReport from '../Reports/DownloadCVEsReport';
-
-jest.useFakeTimers('modern').setSystemTime(new Date('2020-01-01').getTime());
 
 jest.mock("react-redux", () => ({
     ...jest.requireActual("react-redux"),
@@ -25,7 +22,6 @@ const customMiddleWare = store => next => action => {
 
 const mockStore = configureStore([customMiddleWare]);
 const store = mockStore({});
-window.insights = { loadRemediations: jest.fn(), chrome: { getUserPermissions: jest.fn() } }
 
 let wrapper;
 beforeEach(() => {
