+++ src/Components/SmartComponents/Reports/Executive/FirstPage.test.js
@@ -6,8 +6,6 @@ import configureStore from 'redux-mock-store';
 import { Provider } from 'react-redux';
 import { intl } from '../../../../Utilities/IntlProvider';
 
-jest.useFakeTimers('modern').setSystemTime(new Date('2020-01-01').getTime());
-
 const mockStore = configureStore([store => next => action => {}]);
 let store = mockStore({ });
