+++ src/Components/SmartComponents/Reports/ReportsPage.test.js
@@ -17,8 +17,6 @@ jest.mock("react-redux", () => ({
     useSelector: jest.fn()
 }));
 
-jest.useFakeTimers('modern').setSystemTime(new Date('2020-01-01').getTime());
-
 const state = {
     parameters: {}
 }
@@ -31,8 +29,8 @@ const customMiddleWare = store => next => action => {
 }
 
 const mockStore = configureStore([customMiddleWare]);
+
 const store = mockStore(state);
-window.insights = { loadRemediations: jest.fn(), chrome: { getUserPermissions: jest.fn() } };
 
 let wrapper;
 beforeEach(() => {
