+++ src/Components/SmartComponents/SystemCves/SystemCves.test.js
@@ -104,7 +104,6 @@ const customMiddleWare = store => next => action => {
 }
 const mockStore = configureStore([customMiddleWare]);
 const store = mockStore(initialState);
-window.insights = { loadRemediations: jest.fn(), chrome: { getUserPermissions: jest.fn() } }
 
 let wrapper;
 beforeEach(() => {
