+++ src/Components/SmartComponents/SystemCves/SystemCvesTableToolbar.test.js
@@ -14,15 +14,12 @@ import {
 } from '../../../Helpers/TableToolbarHelper';
 import { SYSTEM_DETAILS_HEADER } from '../../../Helpers/constants';
 
-
-jest.mock('../../../Helpers/TableToolbarHelper', () => (
-    {
-        ...(jest.requireActual('../../../Helpers/TableToolbarHelper')),
-        handleChangePage: jest.fn(),
-        handleSetPageSize: jest.fn(),
-        removeFilters: jest.fn()
-    }
-));
+jest.mock('../../../Helpers/TableToolbarHelper', () => ({
+    ...jest.requireActual('../../../Helpers/TableToolbarHelper'),
+    handleChangePage: jest.fn(),
+    handleSetPageSize: jest.fn(),
+    removeFilters: jest.fn()
+}));
 
 const mockContext = {
     cves: createCveListBySystem('CVE-2019-6454', {
@@ -100,8 +97,6 @@ const mockContext = {
         })
     }
 };
-
-window.insights = { loadRemediations: jest.fn(), chrome: { getUserPermissions: jest.fn() } }
 
 const customMiddleWare = store => next => action => {
     next(action);
