+++ src/Components/SmartComponents/SystemDetailsPage/SystemDetails.test.js
@@ -1,5 +1,4 @@
 import SystemDetails from "./SystemDetails";
-import toJson from 'enzyme-to-json';
 import { mountWithIntl } from '../../../Helpers/MiscHelper';
 import { BrowserRouter as Router } from 'react-router-dom';
 import configureStore from 'redux-mock-store';
