+++ src/Components/SmartComponents/SystemsExposedTable/SystemsExposedTable.test.js
@@ -10,6 +10,7 @@ import { act } from "react-dom/test-utils";
 import DownloadReport from '../../../Helpers/DownloadReport';
 
 jest.spyOn(React, 'createRef').mockReturnValueOnce({ current: false });
+
 jest.mock("react-redux", () => ({
     ...jest.requireActual("react-redux"),
     useSelector: jest.fn()
@@ -33,17 +34,6 @@ const customMiddleWare = store => next => action => {
 }
 const mockStore = configureStore([customMiddleWare]);
 let store = mockStore(initialState);
-
-window.insights = {
-    loadInventory: jest.fn(() => Promise.resolve({
-        inventoryConnector: (args) => {
-            const InventoryTable = () => <div>A mock passed!</div>;
-            return ({ InventoryTable });
-        },
-        mergeWithEntities: () => { },
-        mergeWithDetail: () => { }
-    }))
-};
 
 let wrapper;
 beforeEach(() => {
