+++ src/Components/SmartComponents/SystemsPage/SystemTableToolbar.test.js
@@ -13,8 +13,6 @@ const applyMock = jest.fn();
 const handleSelectMock = jest.fn();
 const doOptOutMock = jest.fn();
 
-window.insights = {};
-
 const systems = {
     "data": [
         {
