+++ src/Components/SmartComponents/SystemsPage/SystemsPage.test.js
@@ -1,10 +1,10 @@
 import SystemsPage from './SystemsPage';
 
-describe('Systems', () => {
+describe('SystemsPage', () => {
     let fetchData = jest.fn();
     let systems = { data: [], meta: {} };
 
-    it('fetchData gets called', () => {
+    it('should call fetchData function', () => {
         const systemsPage = <SystemsPage systems={systems} fetchData={fetchData}></SystemsPage>;
         expect(fetchData).toBeCalled;
     });
